Add Exception

+++ lib/qingcloud/cli/utility/file_manager.rb
@@ -1,4 +1,5 @@
 require 'fileutils'
+require 'pathname'
 
 module QingCloud
     module Cli
@@ -18,20 +19,22 @@ module QingCloud
                     FileUtils.mkdir_p directory
                 end
 
+                def read_config_file
+                    read_file Contract::CONFIG_FILE_PATH
+                end
+
+                def write_config_file(content)
+                    write_file Contract::CONFIG_FILE_PATH, content
+                end
+
+                private
+
                 def read_file(file_path)
-                    File.open(file_path) {|file| return file.read }
+                    File.open(file_path) {|file| return file.read } if File.exist? file_path
                 end
 
                 def write_file(file_path, content)
                     File.open(file_path, 'w') { |file| file.write content }
-                end
-
-                def read_config_file
-                    self.read_file Contract::CONFIG_FILE_PATH
-                end
-
-                def write_config_file(content)
-                    self.write_file Contract::CONFIG_FILE_PATH, content
                 end
             end
         end
